PubSub/requirements.txt
google-cloud-pubsub==2.2.0
gunicorn==20.0.4
scipy
oauth2client==3.0.0 
functions-framework==3.4.0
google-cloud-bigquery==3.11.0
